// src/entities/types.ts
export type UserRole = "SUPERADMIN" | "ADMIN" | "MARKETING" | "INVENTORY" | "TEKNISI_LAPANGAN";

export interface AppUser {
    uid: string;
    email: string;
    name: string;
    role: UserRole;
    companyId: string;
    avatar?: string;
    isActive: boolean;
    wa?: string;
    jabatan?: string;
    nip?: string | null;
}
// ...
export type ServiceType = 'GENERAL_PEST_CONTROL' | 'TERMITE_CONTROL' | 'FUMIGATION' | 'RODENT_CONTROL' | 'MOSQUITO_CONTROL' | 'BIRD_CONTROL' | 'BED_BUG_CONTROL' | 'DISINFECTION';
// ...
export interface SPK {
    id: string;
    quotationId: string;
    quotationNoSurat: string;
    customerId: string;
    customerName: string;
    technicianId: string;
    technicianName: string;
    scheduleDate: Date;
    actualStart?: Date;
    actualEnd?: Date;
    durationMin?: number;
    status: "assigned" | "in_progress" | "done";
    gpsStart?: { lat: number; lng: number };
    gpsEnd?: { lat: number; lng: number };
    serviceType: ServiceType;
    perihal: string;
    lokasi: string;
    notes: string;
    alatPekerjaan?: string[];
    companyId: string;
    createdAt: Date;
    createdBy: string;
    createdByName: string;
}